完善聊天功能(含歷史紀錄)
