+++ routes/index.js
@@ -9,7 +9,7 @@ const api = require('./modules/api')
 const userController = require('../controllers/user-controller')
 const { authenticated } = require('../middleware/auth')
 const admin = require('./modules/admin')
-const chat = require('./modules/chat')
+const message = require('./modules/message')
 
 router.use('/api', authenticated, api)
 router.use('/users', authenticated, user)
@@ -23,7 +23,7 @@ router.post('/signup', userController.signUp)
 router.get('/logout', userController.logout)
 router.get('/users/setting/:id', authenticated, userController.getSetting)
 router.put('/users/setting/:id', authenticated, userController.putSetting)
-router.use('/chat', authenticated, chat)
+router.use('/messages', authenticated, message)
 router.use('/', authenticated, (req, res) => res.redirect('/tweets'))
 router.use('/', generalErrorHandler)
 module.exports = router
