--- routes/modules/chat.js
@@ -1,6 +1,0 @@
-const express = require('express')
-const router = express.Router()
-const chatController = require('../../controllers/chat-controller')
-
-router.get('/', chatController.startChatting)
-module.exports = router
