+++ views/layouts/main.hbs
@@ -9,6 +9,7 @@
         <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
         <link rel="stylesheet" type="text/css" href="/css/twitter.css">
         <script src="https://kit.fontawesome.com/c35e0f2e49.js" crossorigin="anonymous"></script>
+        <script src='https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js'></script>
     </head>
 
     <body data-user-id="{{logInUser.id}}">
@@ -19,7 +20,6 @@
     </body>
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
-    <script src='https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js'></script>
     <script src="/javascripts/ajax.js"></script>
     
 </html>
